Repair non provide value bug in position

+++ engine/player.js
@@ -41,10 +41,10 @@ export default class Player {
     }
     updatePosition() {
         if (this.dom !== undefined) {
-            const nodeHeight = parseFloat(getComputedStyle(this.dom).height);
-            const nodeWidth = parseFloat(getComputedStyle(this.dom).width);
-            this.xPosition = parseFloat(getComputedStyle(this.dom).left) + (nodeWidth / 2);
-            this.yPosition = parseFloat(getComputedStyle(this.dom).top) + (nodeHeight / 2);
+            const nodeHeight = parseInt(getComputedStyle(this.dom).height);
+            const nodeWidth = parseInt(getComputedStyle(this.dom).width);
+            this.xPosition = parseInt(getComputedStyle(this.dom).left) + (nodeWidth / 2);
+            this.yPosition = parseInt(getComputedStyle(this.dom).top) + (nodeHeight / 2);
         }
     }
     getPosition() {
